feat: move rest api to newly created rest package
